07/07 12:11

+++ DataWeight.java
@@ -1,9 +1,9 @@
 /**********************************************/
-/*author:金　東柱 6/25更新
+/*author:金　東柱 7/5更新
 /*		 佐野　渉 6/28更新
 /*C4:重み処理部所属
 /*EventWeight:
-/*重み設定画面のイベント処理を記述したクラス
+/*重み入力画面のデータ処理を記述したクラス
 /**********************************************/
 package application;
 
@@ -11,7 +11,14 @@ import java.util.ArrayList;
 import java.util.List;
 
 class DataWeight {
-	void weightSave(List<Integer> wdata)
+	
+	//-------------------------------------------- 
+	//void weightSave(List<Double> wdata)
+	//入力されたデータをユーザ情報管理部に
+	//書込みをする
+	//wdata:入力された重みのリスト
+	//--------------------------------------------
+	void weightSave(List<Double> wdata)
 	{
 		List<Double> weight = new ArrayList<Double>();
 		for(int i = 0; i < wdata.size(); ++i)
@@ -22,13 +29,20 @@ class DataWeight {
 		new UserData().weightWrite(weight);
 	}
 	
+
+	//-------------------------------------------- 
+	//boolean getBoot()
 	//初期起動状態を取得する
+	//--------------------------------------------
 	boolean getBoot()
 	{
 		return new UserData().checkBoot();
 	}
 	
+	//-------------------------------------------- 
+	//void onBoot()
 	//既起動にする
+	//--------------------------------------------
 	void onBoot()
 	{
 		new UserData().writeBoot(true);
